requirements/prod.txt
SQLAlchemy
Flask_Caching
Flask_SQLAlchemy
Flask_Bcrypt
Flask-Migrate
dynaconf
Flask-RESTful
Flask-RESTPlus
